Isolating header in a sub-module
It was getting too large and harder to track.
